fix: remove unused variable assignment and improve error handling in set_proxy function

+++ src-tauri/src/lib.rs
@@ -1,5 +1,3 @@
-#[cfg(target_os = "windows")]
-use png;
 use std::fs;
 use tauri::{AppHandle, Manager, Window, WindowEvent};
 use tauri_plugin_http::reqwest;
@@ -57,7 +55,7 @@ fn get_tray_icon(app: AppHandle) -> Vec<u8> {
         // 将 RGBA 数据转换为 PNG 格式
         let mut png_data = Vec::new();
         {
-            let mut encoder = png::Encoder::new(&mut png_data, width as u32, height as u32);
+            let mut encoder = png::Encoder::new(&mut png_data, width, height);
             encoder.set_color(png::ColorType::Rgba);
             encoder.set_depth(png::BitDepth::Eight);
             let mut writer = encoder.write_header().unwrap();
